v-1.0

+++ src/main/java/com/carolin_violet/travel_system/service/impl/PhotosServiceImpl.java
@@ -40,27 +40,30 @@ public class PhotosServiceImpl extends ServiceImpl<PhotosMapper, Photos> impleme
         wrapper.eq("mark_id", id);
         List<Photos> photoObjectList = baseMapper.selectList(wrapper);
 
-        // 遍历图片列表，切割字符串，将所有图片路径存入keys
-        List<String> keys = new ArrayList<>();
-        for (Photos photo : photoObjectList) {
-            String[] arr = photo.getPicture().split("/");
-            String url = arr[arr.length-2] + "/" + arr[arr.length-1];
-            keys.add(url);
+        if (photoObjectList.size() > 0) {
+            // 遍历图片列表，切割字符串，将所有图片路径存入keys
+            List<String> keys = new ArrayList<>();
+            for (Photos photo : photoObjectList) {
+                String[] arr = photo.getPicture().split("/");
+                String url = arr[arr.length-2] + "/" + arr[arr.length-1];
+                keys.add(url);
+            }
+
+            // 创建OSSClient实例。
+            OSS ossClient = new OSSClientBuilder().build(endpoint, accessKeyId, accessKeySecret);
+            // 删除请求
+            DeleteObjectsRequest deleteRequest = new DeleteObjectsRequest(bucketName);
+            // 赋值需要删除的文件
+            deleteRequest.setKeys(keys);
+            //deleteRequest.setEncodingType(URL_ENCODING);
+            // 调用删除
+            DeleteObjectsResult deleteObjectsResult = ossClient.deleteObjects(deleteRequest);
+            // 获取删除的文件
+            List<String> deletedObjects = deleteObjectsResult.getDeletedObjects();
+            // 关闭OSSClient。
+            ossClient.shutdown();
         }
 
-        // 创建OSSClient实例。
-        OSS ossClient = new OSSClientBuilder().build(endpoint, accessKeyId, accessKeySecret);
-        // 删除请求
-        DeleteObjectsRequest deleteRequest = new DeleteObjectsRequest(bucketName);
-        // 赋值需要删除的文件
-        deleteRequest.setKeys(keys);
-        //deleteRequest.setEncodingType(URL_ENCODING);
-        // 调用删除
-        DeleteObjectsResult deleteObjectsResult = ossClient.deleteObjects(deleteRequest);
-        // 获取删除的文件
-        List<String> deletedObjects = deleteObjectsResult.getDeletedObjects();
-        // 关闭OSSClient。
-        ossClient.shutdown();
 
         // 2. 删除表中图片信息
         baseMapper.delete(wrapper);
